requirements.txt
streamlit
sentence-transformers
PyPDF2
python-docx
transformers
torch
requests
beautifulsoup4
